MTP-1210: renamed to menuPointsIcons
